requirements.txt
opencv-python>=4.8.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
pyzbar>=0.1.9
pyinstaller>=6.0.0
streamlit>=1.28.0
